feat: Add lubridate package for date manipulation
The code changes in `vertical_bar_chart.R` import the `lubridate` package to handle date manipulation. This allows for converting the date and time values to a weekly grouping format using the `floor_date` function. This improves the accuracy of the vertical bar chart module by grouping the data by week instead of the original date and time values.